--- packages/design-tokens/src/types/types.ts
@@ -1,5 +1,0 @@
-type InternalPlatform = "docs";
-
-export type Platform = "css" | "scss" | "es6" | InternalPlatform;
-export type Stylesheet = Extract<Platform, "css" | "scss">;
-export type RegisterFn = () => void;
